2021/12/23 第一次发布到自己的分支上

+++ TSM/src/main/java/com/tsm/entity/Studentfiles.java
@@ -65,7 +65,7 @@ public class Studentfiles implements Serializable {
 
     @TableField("DELETED")
     @TableLogic
-    private BigDecimal deleted;
+    private int deleted;
 
 
 }
